requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pinecone==6.0.2
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
PyPDF2==3.0.1
pydantic[email]==2.5.0
python-multipart==0.0.6
numpy==1.24.3
groq==0.4.1
httpx==0.27.2
